create seprate class for graphql error
