+++ lib/api/handlers/graphqlWithoutUserClaims.ts
@@ -1,56 +1,9 @@
-import {
-  fetcher,
-  getAdditionalHeader,
-  getOperationDetails,
-  getUserClaimsFromRequest,
-} from '../util'
+import { fetcher, getAdditionalHeader, getOperationDetails } from '../util'
+import { GraphQLError } from '../util/graphql-error-class'
 import { KIBO_HEADERS } from '@/lib/constants'
 import { NextApiRequestWithLogger } from '@/lib/types'
 
 import type { NextApiResponse } from 'next'
-
-type GraphQLOperation = {
-  operationType: string | undefined
-  operationName: string | undefined
-}
-
-class GraphQLError extends Error {
-  metadata: GraphQLOperation
-  rawErrors: any
-  code: string | undefined
-  correlationId?: string | null
-
-  constructor(
-    responseErrors: any,
-    operationMetaData: GraphQLOperation,
-    correlationId?: string | null
-  ) {
-    super(responseErrors[0]?.extensions?.response?.body?.message)
-
-    this.name = 'GraphQLError' // Set the name of the error
-    this.code = responseErrors[0]?.extensions.response.status
-    this.rawErrors = responseErrors
-    this.metadata = operationMetaData
-    this.correlationId = correlationId
-  }
-
-  toJson() {
-    return {
-      gql: {
-        ...this.metadata,
-        statusCode: this.code,
-      },
-      correlationId: this.correlationId,
-    }
-  }
-
-  dumpErrors(): string {
-    if (this.rawErrors && typeof this.rawErrors === 'object') {
-      return JSON.stringify(this.rawErrors, null, 2)
-    }
-    return ''
-  }
-}
 
 export default async function graphQLWithoutUserClaimsHandler(
   req: NextApiRequestWithLogger,
@@ -61,7 +14,6 @@ export default async function graphQLWithoutUserClaimsHandler(
     const gqlDetails = getOperationDetails(query)
     req.logger.info('incoming graphql request', { gql: gqlDetails })
 
-    const userClaims = await getUserClaimsFromRequest(req, res)
     const headers = getAdditionalHeader(req)
     const response = await fetcher({ query, variables }, { headers })
